requirements.txt
torchvision
scikit-learn
Pillow
opencv-python
numpy
matplotlib
python-dotenv